refactor: restructure preDelete to match postCreate architecture
- Move executePreDeleteCommands to dedicated pre-delete.ts file
- Update deleteWorktree to accept preDeleteCommands parameter from context
- Remove inline config loading from delete.ts
- Update CLI handler to pass config from context
- Add pre-delete module to core exports
- Update all tests to match new structure

This ensures consistency between preDelete and postCreate implementations,
following the same architectural patterns.
