+++ packages/cli/src/handlers/delete.ts
@@ -94,6 +94,7 @@ export async function deleteHandler(args: string[]): Promise<void> {
       {
         force: forceDelete,
       },
+      context.config?.preDelete?.commands,
     );
 
     if (isErr(result)) {
