+++ packages/core/src/index.ts
@@ -12,5 +12,6 @@ export * from "./worktree/validate.ts";
 export * from "./worktree/select.ts";
 export * from "./worktree/file-copier.ts";
 export * from "./worktree/post-create.ts";
+export * from "./worktree/pre-delete.ts";
 export * from "./exec.ts";
 export * from "./shell.ts";
