+++ packages/mcp/src/tools/delete-worktree.ts
@@ -26,6 +26,7 @@ export const deleteWorktreeTool: Tool<typeof schema> = {
       {
         force,
       },
+      context.config?.preDelete?.commands,
     );
 
     if (!isOk(result)) {
